requirements.txt
mcp
fastmcp
selenium
langchain-groq
langgraph
langchain-mcp-adapters
python-dotenv
langchain-ollama
langchain_core